back_end/requirements.txt
asgiref==3.8.1
Django==5.1.6
djangorestframework==3.15.2
sqlparse==0.5.3
typing_extensions==4.12.2
djoser==2.3.1
django-cors-headers==4.3.1